- Fixed huge amount of warnings with MS compilers  - Updates for both VC6 and VC.net IDE project files
git-svn-id: file:///fltk/svn/trunk@4787 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ OpenGL/gl_draw.cxx
@@ -230,7 +230,7 @@ void fltk::gldrawimage(const uchar* b, int x, int y, int w, int h, int d, int ld
 void glWindowPos2i(int x, int y) {
   if (x < 0 || y < 0) {
     glRasterPos2i(0,0);
-    glBitmap(0,0,0,0,x,y,0);
+    glBitmap(0,0,0,0,(GLfloat)x,(GLfloat)y,0);
   } else {
     glRasterPos2i(x,y);
   }
